Introduce alternate way of using RSuite in Next.js
Newer developers might run into problems trying to use RSuite in Next.js. This edit makes it more developer friendly and usable. I believe that the simple usage (compiled less) should be mentioned and will be preferred by some developers (I personally was not aware of it and spent a lot of time tuning Webpack and less-loader to get what I want).

This is just a simple suggestion from me to make the docs better.